Fix bug and update seeding with reusable code for test seeding and dev seeding.
