--- Infrastructure/Seeds/Dev/SeedConstantDev.cs
@@ -1,6 +1,0 @@
-namespace Infrastructure.Seeds.Dev;
-
-public static class SeedConstantDev
-{
-    public const string general = "General";
-}
