+++ Infrastructure/Seeds/SeedConstant.cs
@@ -1,0 +1,6 @@
+namespace Infrastructure.Seeds;
+
+public static class SeedConstant
+{
+    public const string general = "General";
+}
