+++ Infrastructure/Seeds/SeedException.cs
@@ -1,0 +1,6 @@
+namespace Infrastructure.Seeds;
+
+public class SeedException(
+    string message = "An unexpected error occurred during seeding."
+) : Exception(message)
+{ }
